% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_pkg_registry_files}
\alias{check_pkg_registry_files}
\title{Check Paths in Registry Files}
\usage{
check_pkg_registry_files(pkg = system.file(package = "RcppCWB"),
  set = FALSE)
}
\arguments{
\item{pkg}{Full path to package directory}

\item{set}{Logical, whether}
}
\value{
Logical value, whether home directories are set correctly.
}
\description{
Check Paths in Registry Files
}
