% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl_functions}
\alias{cl_functions}
\alias{cl_attribute_size}
\alias{cl_lexicon_size}
\alias{cl_cpos2struc}
\alias{cl_cpos2str}
\alias{cl_cpos2id}
\alias{cl_struc2cpos}
\alias{cl_id2str}
\alias{cl_struc2str}
\alias{cl_regex2id}
\alias{cl_str2id}
\alias{cl_id2freq}
\alias{cl_id2cpos}
\alias{cl_cpos2lbound}
\alias{cl_cpos2rbound}
\title{Exposed functions of the corpus library (CL).}
\usage{
cl_attribute_size(corpus, attribute, attribute_type,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cl_lexicon_size(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"))

cl_cpos2struc(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cl_cpos2str(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  cpos)

cl_cpos2id(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  cpos)

cl_struc2cpos(corpus, s_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  struc)

cl_id2str(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), id)

cl_struc2str(corpus, s_attribute, struc,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cl_regex2id(corpus, p_attribute, regex,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cl_str2id(corpus, p_attribute, str, registry = Sys.getenv("CORPUS_REGISTRY"))

cl_id2freq(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY"))

cl_id2cpos(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY"))

cl_cpos2lbound(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cl_cpos2rbound(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{name of a CWB corpus (upper case)}

\item{attribute}{name of a s- or p-attribute}

\item{attribute_type}{either "p" or "s"}

\item{registry}{path to the registry directory}

\item{p_attribute}{a p-attribute}

\item{s_attribute}{a s-attribute}

\item{cpos}{corpus positions (integer vector)}

\item{struc}{strucs (integer vector)}

\item{id}{id of a token}

\item{regex}{a regular expression}

\item{str}{a character string}
}
\description{
Exposed functions of the corpus library (CL).
}
