% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl_corpora}
\alias{cl_corpora}
\title{Show CL corpora}
\usage{
cl_corpora()
}
\value{
\code{TRUE} if corpus could be loaded and \code{FALSE} if not.
}
\description{
Show CL corpora
}
\examples{
cl_load_corpus("REUTERS")
}
