% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCWB_package.R
\name{cl_functions}
\alias{cl_functions}
\alias{attribute_size}
\alias{lexicon_size}
\alias{cpos2struc}
\alias{cpos2str}
\alias{cpos2id}
\alias{struc2cpos}
\alias{id2str}
\alias{struc2str}
\alias{regex2id}
\alias{str2id}
\alias{id2freq}
\alias{id2cpos}
\alias{cpos2lbound}
\alias{cpos2rbound}
\title{Functions of the Corpus Library (CL).}
\arguments{
\item{corpus}{name of a CWB corpus (upper case)}

\item{attribute}{name of a s- or p-attribute}

\item{attribute_type}{either "p" or "s"}

\item{registry}{path to the registry directory}

\item{p_attribute}{a p-attribute}

\item{s_attribute}{a s-attribute}

\item{cpos}{corpus positions (integer vector)}

\item{struc}{strucs (integer vector)}

\item{id}{id of a token}

\item{regex}{a regular expression}

\item{str}{a character string}
}
\description{
Functions of the Corpus Library (CL).
}
\section{Functions}{

\describe{
  \item{\code{attribute_size(corpus, attribute, attribute_type, registry)}}{}
  \item{\code{lexicon_size(corpus, p_attribute, registry)}}{}
  \item{\code{cpos2struc(corpus, s_attribute, cpos, registry)}}{}
  \item{\code{cpos2str(corpus, p_attribute, registry, cpos)}}{}
  \item{\code{cpos2id(corpus, p_attribute, registry, cpos)}}{}
  \item{\code{struc2cpos(corpus, s_attribute, registry, struc)}}{}
  \item{\code{id2str(corpus, p_attribute, registry, id)}}{}
  \item{\code{struc2str(corpus, s_attribute, struc, registry)}}{}
  \item{\code{regex2id(corpus, p_attribute, regex, registry)}}{}
  \item{\code{str2id(corpus, p_attribute, str, registry)}}{}
  \item{\code{id2freq(corpus, p_attribute, id, registry)}}{}
  \item{\code{id2cpos(corpus, p_attribute, id, registry)}}{}
  \item{\code{cpos2lbound(corpus, s_attribute, cpos, registry)}}{}
  \item{\code{cpos2rbound(corpus, s_attribute, cpos, registry)}}{}
}
}

