% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl_load_corpus}
\alias{cl_load_corpus}
\title{Load corpus}
\usage{
cl_load_corpus(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{A length-one \code{character} vector with the corpus ID.}

\item{registry}{A length-one \code{character} vector with the registry directory.}
}
\value{
\code{TRUE} if corpus could be loaded and \code{FALSE} if not.
}
\description{
Load corpus
}
\examples{
cl_load_corpus("REUTERS")
}
