% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R, R/RcppExports.R
\name{cl_rework}
\alias{cl_rework}
\alias{s_attr}
\alias{p_attr}
\alias{p_attr_size}
\alias{s_attr_size}
\alias{p_attr_lexicon_size}
\alias{cpos_to_struc}
\alias{cpos_to_str}
\alias{cpos_to_id}
\alias{struc_to_cpos}
\alias{struc_to_str}
\alias{regex_to_id}
\alias{str_to_id}
\alias{id_to_freq}
\alias{id_to_cpos}
\alias{cpos_to_lbound}
\alias{cpos_to_rbound}
\title{Experimental low-level CL access.}
\usage{
s_attr(corpus, s_attribute, registry)

p_attr(corpus, p_attribute, registry)

p_attr_size(p_attr)

s_attr_size(s_attr)

p_attr_lexicon_size(p_attr)

cpos_to_struc(s_attr, cpos)

cpos_to_str(p_attr, cpos)

cpos_to_id(p_attr, cpos)

struc_to_cpos(s_attr, struc)

struc_to_str(s_attr, struc)

regex_to_id(p_attr, regex)

str_to_id(p_attr, str)

id_to_freq(p_attr, id)

id_to_cpos(p_attr, id)

cpos_to_lbound(s_attr, cpos)

cpos_to_rbound(s_attr, cpos)
}
\arguments{
\item{corpus}{ID of a CWB corpus (length-one \code{character} vector).}

\item{s_attribute}{A structural attribute (length-one \code{character} vector).}

\item{registry}{Registry directory.}

\item{p_attribute}{A positional attribute (length-one \code{character} vector).}

\item{p_attr}{A \code{externalptr} referencing a p-attribute.}

\item{s_attr}{A \code{externalptr} referencing a p-attribute.}

\item{cpos}{An \code{integer} vector of corpus positions.}

\item{struc}{A length-one \code{integer} vector with a struc.}

\item{regex}{A regular expression.}

\item{str}{A \code{character} vector.}

\item{id}{An \code{integer} vector with token ids.}
}
\description{
Set of functions with same functionality as cl_* functions to improve the
ease of writing code.
}
