% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{corpus_data_dir}
\alias{corpus_data_dir}
\title{Get data directory of a corpus}
\usage{
corpus_data_dir(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{A length-one \code{character} vector with the corpus ID.}

\item{registry}{A length-one \code{character} vector with the registry directory.}
}
\value{
A length-one \code{character} vector stating the data directory.
}
\description{
Extract the data directory from the intenal C representation of the content
of the registry file for a corpus.
}
\examples{
corpus_data_dir("REUTERS")
}
