% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{corpus_is_loaded}
\alias{corpus_is_loaded}
\title{Check whether corpus is loaded}
\usage{
corpus_is_loaded(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{A length-one \code{character} vector with the corpus ID.}

\item{registry}{A length-one \code{character} vector with the registry directory.}
}
\value{
\code{TRUE} if corpus is loaded and \code{FALSE} if not.
}
\description{
Check whether corpus is loaded
}
