% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cqp_query}
\alias{cqp_query}
\title{Run a CQP query.}
\usage{
cqp_query(inMother, inChild, inQuery)
}
\arguments{
\item{inMother}{the corpus}

\item{inChild}{the temporary corpus}

\item{inQuery}{the query}
}
\description{
Run a CQP query.
}
