% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{cqp_query}
\alias{cqp_query}
\alias{cqp_dump_subcorpus}
\alias{cqp_subcorpus_size}
\alias{cqp_list_subcorpora}
\title{Perform a CQP Query.}
\usage{
cqp_query(corpus, query)

cqp_dump_subcorpus(corpus, child = "FOO")

cqp_subcorpus_size(corpus, child = "FOO")

cqp_list_subcorpora(corpus)
}
\arguments{
\item{corpus}{a CWB corpus}

\item{query}{}
}
\description{
Perform a CQP Query.
}
