% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{cwb_list_corpora}
\alias{cwb_list_corpora}
\title{List Available CWB Corpora.}
\usage{
cwb_list_corpora()
}
\description{
List Available CWB Corpora.
}
\examples{
\donttest{
registry <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
Sys.setenv("CORPUS_REGISTRY" = registry)
# initialize_cqp()
cwb_list_corpora()
}
}
