% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb_makeall}
\alias{cwb_makeall}
\title{CWB Utility Functions}
\usage{
cwb_makeall(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{name of a CWB corpus (upper case)}

\item{p_attribute}{name p-attribute}

\item{registry}{path to the registry directory, defaults to the value of the
environment variable CORPUS_REGISTRY}
}
\description{
Wrappers for CWB utilities (cwb-makeall, cwb-huffcode, cwb-compress-rdx).
}
