% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb_version}
\alias{cwb_version}
\title{Get CWB version}
\usage{
cwb_version()
}
\value{
A \code{numeric_version} vector.
}
\description{
Get the CWB version used and available when compiling the source code.
}
\examples{
cwb_version()
cwb_version() > numeric_version("3.4.14")
}
