% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_functions.R
\name{decode_s_attribute}
\alias{decode_s_attribute}
\title{Decode Structural Attribute.}
\usage{
decode_s_attribute(corpus, s_attribute,
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{a CWB corpus}

\item{s_attribute}{a structural attribute}

\item{registry}{registry directory}
}
\value{
a character vector
}
\description{
Return values for all strucs of a structural attribute.
}
\examples{
registry <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
decode_s_attribute(corpus = "REUTERS", s_attribute = "id", registry = registry)
Sys.setenv(CORPUS_REGISTRY = registry)
decode_s_attribute(
  corpus = "REUTERS", s_attribute = "places",
  registry = registry
  )
}
