% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\name{decodeSAttribute}
\alias{decodeSAttribute}
\alias{decode_s_attribute}
\title{Decode s-attributes}
\usage{
decodeSAttribute(corpus, sAttribute, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{character}

\item{sAttribute}{character}

\item{registry}{character}
}
\description{
Proof of concept - not faster than rcqp::cqi_struc2str.
}
