% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_pkg_registry}
\alias{get_pkg_registry}
\title{Get Registry Directory Within Package}
\usage{
get_pkg_registry(pkgname = "RcppCWB")
}
\arguments{
\item{pkgname}{Name of package (character vector)}
}
\description{
Get Registry Directory Within Package
}
