% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCWB_package.R
\name{getCbowMatrix}
\alias{getCbowMatrix}
\alias{get_cbow_matrix}
\title{Get matrix with CBOW.}
\usage{
getCbowMatrix(corpus, pAttribute = "word",
  registry = Sys.getenv("CORPUS_REGISTRY"), matrix, window)
}
\arguments{
\item{corpus}{character}

\item{pAttribute}{character}

\item{registry}{character}

\item{matrix}{a matrix with regions}

\item{window}{intger, window size}
}
\description{
Substantially faster than lapply(cqi_struc2cpos).
}
\examples{
Sys.setenv(CORPUS_REGISTRY = system.file(package = "RcppCWB", "extdata", "cwb", "registry"))
getCbowMatrix("REUTERS", pAttribute = "word", matrix = matrix(c(1,10), nrow = 1), window = 5L)
}
