% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCWB_package.R
\name{getCountVector}
\alias{getCountVector}
\alias{get_count_vector}
\title{Count tokens in corpus.}
\usage{
getCountVector(corpus, pAttribute = "word",
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{character}

\item{pAttribute}{character}

\item{registry}{character}
}
\description{
Substantially faster than cqi_cpos2str/tabulate.
}
\examples{
Sys.setenv(CORPUS_REGISTRY = system.file(package = "RcppCWB", "extdata", "cwb", "registry"))
getCountVector("REUTERS", pAttribute = "word")
}
