% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\name{getRegionMatrix}
\alias{getRegionMatrix}
\alias{get_region_matrix}
\title{Get Matrix with Regions.}
\usage{
getRegionMatrix(corpus, sAttribute = "word", strucs,
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{character}

\item{sAttribute}{character}

\item{strucs}{integer}

\item{registry}{character}
}
\description{
Substantially faster than lapply(cqi_struc2cpos).
}
