% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{ids_to_count_matrix}
\alias{ids_to_count_matrix}
\title{Perform Count for Vector of IDs.}
\usage{
ids_to_count_matrix(ids)
}
\arguments{
\item{ids}{a vector of ids (integer values)}
}
\description{
The return value is a two-column integer matrix. Column one represents the
unique ids of the input vector, column two the respective number of
occurrences / counts.
}
\examples{
ids <- c(1L, 5L, 5L, 7L, 7L, 7L, 7L)
ids_to_count_matrix(ids)
table(ids) # alternative to get a similar result
}
