% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{initialize_cwb}
\alias{initialize_cwb}
\alias{get_registry}
\title{Initialize CQP}
\usage{
initialize_cwb()

get_registry()
}
\description{
Initialize CQP

Get Registry Directory
}
\examples{
initialize_cwb()
get_registry()
get_progname()
}
