% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{initialize_cqp}
\alias{initialize_cqp}
\alias{get_cqp_registry}
\alias{set_cqp_registry}
\title{Initialize Corpus Query Processor (CQP).}
\usage{
initialize_cqp()

get_cqp_registry()

set_cqp_registry(registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{registry}{the registry directory}
}
\description{
CQP will need to know where to look for CWB indexed corpora. Upon
initialization of CQP, the registry directory is set, by calling the function
\code{initialize_cqp} during the startup procedure. To reset the registry,
use the function \code{set_cqp_registry}. To get the registry used by CQP,
use \code{get_cqp_registry}.
}
\examples{
# initialize_cqp()
get_cqp_registry()
registry_new <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
# set_cqp_registry(registry = registry_new)
}
