% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_attr_default}
\alias{p_attr_default}
\title{Get default p-attribute}
\usage{
p_attr_default()
}
\value{
A length-one \code{character} vector.
}
\description{
Usually the default p-attribute will be "word". Use this function to avoid
a hard-coded solution. Extracts the default attribute defined in the CWB
source code.
}
