% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{drive letters}
\alias{drive letters}
\alias{path_get_drive_letter}
\alias{path_remove_drive_letter}
\alias{path_add_drive_letter}
\title{Drive Letter Handling for Windows.}
\usage{
path_get_drive_letter(path)

path_remove_drive_letter(path)

path_add_drive_letter(path, drive_letter)
}
\arguments{
\item{path}{a path declaration}

\item{drive_letter}{a drive letter}
}
\description{
Auxiliary functions for handling drive letters on Windows systems.
}
\examples{
path_get_drive_letter("D:/PATH/TO/REGISTRY")
path_get_drive_letter("d:/PATH/TO/REGISTRY")
path_remove_drive_letter("D:/PATH/TO/REGISTRY")
path_remove_drive_letter("d:/PATH/TO/REGISTRY")
}
