% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\docType{package}
\name{polmineR.Rcpp}
\alias{polmineR.Rcpp}
\alias{attribute_size}
\alias{lexicon_size}
\alias{cpos2struc}
\alias{cpos2str}
\alias{cpos2id}
\alias{struc2cpos}
\alias{id2str}
\alias{struc2str}
\alias{regex2id}
\alias{str2id}
\alias{id2freq}
\alias{id2cpos}
\alias{cpos2lbound}
\alias{cpos2rbound}
\alias{polmineR.Rcpp-package}
\title{Rcpp extension for polmineR.}
\description{
Based on the functionality of the cl corpus library, this package
offers wrappers for core functions of the corpus library, and 
functions for performance critial tasks in the polmineR package.
}
\references{
http://polmine.sowi.uni-due.de
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
