% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCWB_package.R
\docType{package}
\name{RcppCWB}
\alias{RcppCWB}
\alias{RcppCWB-package}
\title{Rcpp bindings for the corpus library (CL).}
\description{
The corpus library (CL) is a C library offering low-level access 
to CWB-indexed corpora. This package
offers wrappers for core functions of the corpus library, and 
functions for performance critial tasks in the polmineR package.
}
\references{
http://polmine.sowi.uni-due.de
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
