% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_functions.R
\name{regions_to_count_matrix}
\alias{regions_to_count_matrix}
\title{Count for Region Matrix.}
\usage{
regions_to_count_matrix(corpus, p_attribute,
  registry = Sys.getenv("CORPUS_REGISTRY"), matrix)
}
\arguments{
\item{corpus}{a CWB corpus}

\item{p_attribute}{a positional attribute}

\item{registry}{registry directory}

\item{matrix}{a region matrix}
}
\description{
Count tokens in a two-column region matrix with left and right corpus positions.
The function returns a two-column matrix with the token ids (column 1) and the number
of occurrencs / counts (column 2).
}
\examples{
registry <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
m <- get_region_matrix(
  corpus = "REUTERS", s_attribute = "places",
  strucs = 5L:5L, registry = registry
  )
y <- regions_to_count_matrix(
  corpus = "REUTERS", p_attribute = "word",
  registry = registry, matrix = m
  )
df <- as.data.frame(y)
colnames(df) <- c("token_id", "count")
df[["token"]] <- cwb_id2str(
  "REUTERS", p_attribute = "word",
  registry = registry, id = df[["token_id"]]
  )
df[order(df[["count"]], decreasing = TRUE),]
head(df)
}
