% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_functions.R
\name{regions_to_ids}
\alias{regions_to_ids}
\title{Get token IDs for Region Matrix.}
\usage{
regions_to_ids(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  matrix)
}
\arguments{
\item{corpus}{a CWB corpus}

\item{p_attribute}{a positional attribute}

\item{registry}{registry directory}

\item{matrix}{a regions matrix}
}
\description{
Get token IDs for Region Matrix.
}
\examples{
registry <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
m <- get_region_matrix(
  corpus = "REUTERS", s_attribute = "places",
  strucs = 4L:5L, registry = registry
  )
ids <- regions_to_ids(
  corpus = "REUTERS", p_attribute = "word",
  registry = registry, matrix = m
  )
tokenstream <- cwb_id2str(
  corpus = "REUTERS", p_attribute = "word",
  registry = registry, id = ids
  )
txt <- paste(tokenstream, collapse = " ")
txt
}
