% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{region_to_strucs}
\alias{region_to_strucs}
\alias{region_matrix_to_struc_matrix}
\title{Get vector with min and max struc of s-attribute within a region}
\usage{
region_to_strucs(corpus, s_attribute, region, registry = NULL)

region_matrix_to_struc_matrix(
  corpus,
  s_attribute,
  region_matrix,
  registry = NULL
)
}
\arguments{
\item{corpus}{ID of a CWB corpus.}

\item{s_attribute}{Name of nested structural attribute.}

\item{region}{Vector with left and right corpus position of region.}

\item{registry}{Path of the registry directory. If \code{NULL} (default), value
of environment variable 'CORPUS_REGISTRY' will be used.}

\item{region_matrix}{A two-column \code{matrix} with regions, left corpus
positions in column 1, right corpus positions in column 2.}
}
\value{
A length-two integer vector.
}
\description{
Look up the minimum and maximum struc of a s-attribute within a region.
Works for nested s-attributes. If there are no regions of the s-attribute
within the region, a vector with (two) \code{NA} values is returned.
}
