% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\name{regionsToIds}
\alias{regionsToIds}
\alias{regions_to_ids}
\title{Turn regions into ids.}
\usage{
regionsToIds(corpus, pAttribute, registry, matrix)
}
\arguments{
\item{corpus}{character}

\item{pAttribute}{character}

\item{registry}{character}

\item{matrix}{matrix (2 columns) with regions}
}
\description{
Turn regions into ids.
}
