% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{corpus_data_dir}
\alias{corpus_data_dir}
\alias{corpus_info_file}
\title{Get information from registry file}
\usage{
corpus_data_dir(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_info_file(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{A length-one \code{character} vector with the corpus ID.}

\item{registry}{A length-one \code{character} vector with the registry directory.}
}
\value{
A length-one \code{character} vector stating the data directory.
}
\description{
Extract information from the internal C representation of registry data.
}
\details{
\code{corpus_data_dir()} will return the data directory (class \code{fs_path})
where the binary files of a corpus are kept (a directory also known as
'home' directory).

\code{corpus_info_file()} will return the path to the info file for a
corpus (class \code{fs_path} object).
}
\examples{
corpus_data_dir("REUTERS")
corpus_info_file("REUTERS", registry = get_tmp_registry())
}
