% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{registry_write}
\alias{registry_write}
\title{Generate and Write Registry File.}
\usage{
registry_write(registry_dir, corpus, description = "", data_dir,
  info_file = NULL, encoding = c("latin1", "utf-8"), language = c("de",
  "en", "fr"), p_attributes = "word", s_attributes = NULL)
}
\arguments{
\item{registry_dir}{the registry directory}

\item{corpus}{name of the corpus, will be used as ID (using tolower)}

\item{description}{long descriptive name for the corpus}

\item{data_dir}{path to binary data files}

\item{info_file}{optional info file, if NULL, we assume that file ".info.md" is in \code{data_dir}}

\item{encoding}{corpus encoding, should be "latin1" or "utf-8"}

\item{language}{language ISO code, such as "de", "en", "fr"}

\item{p_attributes}{positional attributes to be declared}

\item{s_attributes}{structural attributes to be declared}
}
\value{
the character vector of the registry file is returned invisibly
}
\description{
Generate a registry file that declares positional and structural attributes.
}
\details{
A CWB indexed corpus needs to be described in a registry file. See the CWB
Corpus Encoding Tutorial, Appendix A (pp. 21f) for an example
(\url{http://cwb.sourceforge.net/files/CWB_Encoding_Tutorial.pdf}).
}
