% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{subcorpus_get_ranges}
\alias{subcorpus_get_ranges}
\title{Get ranges of subcorpus}
\usage{
subcorpus_get_ranges(subcorpus_pointer)
}
\arguments{
\item{subcorpus_pointer}{A pointer (class \code{externalptr}) referencing a CWB
subcorpus.}
}
\description{
Get ranges of subcorpus
}
