% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{use_tmp_registry}
\alias{use_tmp_registry}
\title{Use Temporary Registry}
\usage{
use_tmp_registry(pkg = system.file(package = "RcppCWB"))
}
\arguments{
\item{pkg}{full path to a package}
}
\description{
Use a temporary registry directory to describe and access the corpora in 
a package
}
