% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl_functions}
\alias{cl_functions}
\alias{cwb_attribute_size}
\alias{cwb_lexicon_size}
\alias{cwb_cpos2struc}
\alias{cwb_cpos2str}
\alias{cwb_cpos2id}
\alias{cwb_struc2cpos}
\alias{cwb_id2str}
\alias{cwb_struc2str}
\alias{cwb_regex2id}
\alias{cwb_str2id}
\alias{cwb_id2freq}
\alias{cwb_id2cpos}
\alias{cwb_cpos2lbound}
\alias{cwb_cpos2rbound}
\title{Exposed functions of the corpus library (CL).}
\usage{
cwb_attribute_size(corpus, attribute, attribute_type,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_lexicon_size(corpus, p_attribute,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_cpos2struc(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_cpos2str(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  cpos)

cwb_cpos2id(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  cpos)

cwb_struc2cpos(corpus, s_attribute, registry = Sys.getenv("CORPUS_REGISTRY"),
  struc)

cwb_id2str(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), id)

cwb_struc2str(corpus, s_attribute, struc,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_regex2id(corpus, p_attribute, regex,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_str2id(corpus, p_attribute, str, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_id2freq(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_id2cpos(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_cpos2lbound(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_cpos2rbound(corpus, s_attribute, cpos,
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{name of a CWB corpus (upper case)}

\item{attribute}{name of a s- or p-attribute}

\item{attribute_type}{either "p" or "s"}

\item{registry}{path to the registry directory}

\item{p_attribute}{a p-attribute}

\item{s_attribute}{a s-attribute}

\item{cpos}{corpus positions (integer vector)}

\item{struc}{strucs (integer vector)}

\item{id}{id of a token}

\item{regex}{a regular expression}

\item{str}{a character string}
}
\description{
Exposed functions of the corpus library (CL).
}
