% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp.R
\name{cqp_query}
\alias{cqp_query}
\alias{cqp_dump_subcorpus}
\alias{cqp_subcorpus_size}
\alias{cqp_list_subcorpora}
\title{Execute CQP Query and Retrieve Results.}
\usage{
cqp_query(corpus, query)

cqp_dump_subcorpus(corpus, child = "QUERY")

cqp_subcorpus_size(corpus, child = "QUERY")

cqp_list_subcorpora(corpus)
}
\arguments{
\item{corpus}{a CWB corpus}

\item{query}{a CQP query}

\item{child}{a subcorpus}
}
\description{
Using CQP queries requires a two-step procedure: At first, you execute a
query using \code{cqp_query}. Then, \code{cqp_dump_subcorpus} will return a
matrix with the regions of the matches for the query.
}
\details{
The \code{cqp_subcorpus_size} function returns the number of matches for the
CQP query.
}
\examples{
registry <- system.file(package = "RcppCWB", "extdata", "cwb", "registry")
Sys.setenv("CORPUS_REGISTRY" = registry)
# cqp_initialize()
cqp_query(corpus = "REUTERS", query = '"oil";')
cqp_subcorpus_size("REUTERS")
cqp_dump_subcorpus("REUTERS")

cqp_query(corpus = "REUTERS", query = '"crude" "oil";')
cqp_subcorpus_size("REUTERS", child = "QUERY")
cqp_dump_subcorpus("REUTERS")
}
