% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\name{getCountVector}
\alias{getCountVector}
\alias{get_count_vector}
\title{Count tokens in corpus.}
\usage{
getCountVector(corpus, pAttribute = "word",
  registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{character}

\item{pAttribute}{character}

\item{registry}{character}
}
\description{
Substantially faster than cqi_cpos2str/tabulate.
}
