% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.Rcpp_package.R
\name{regionsToCountMatrix}
\alias{regionsToCountMatrix}
\alias{regions_to_count_matrix}
\title{Get matrix with ids and counts.}
\usage{
regionsToCountMatrix(corpus, pAttribute = "word",
  registry = Sys.getenv("CORPUS_REGISTRY"), matrix)
}
\arguments{
\item{corpus}{character}

\item{pAttribute}{character}

\item{registry}{character}

\item{matrix}{a matrix with regions}
}
\description{
Reduces time by one third.
}
